**GIANT HISTORY REWRITE**
Prior history was derived from commit 101727adc286d4d1ab30b4abe2a52d4ac471ead5
of the wwt-web-client repository using the following command:

```
git filter-branch --subdirectory-filter HTML5SDK/wwtlib -- --all
```

This basically takes the history of that repository, pretends that
HTML5SDK/wwtlib/ was the root of it, and rewrites the history to only include
files from that directory. Notably that preserves the modification histories
but throws away the numerous generated files contained in the current history
of wwt-web-client, which make the repo very large.

Going forward, this repository will be slimmed down and contain only the code
needed to generate the WebGL engine. The code for the Webclient web
application will be maintained in the other repository.