Start working on real URL rewriting.
It looks like it's working for pins.png, at least! Surely much will
have to be worked out. Note that we need to move URLHelpers.cs to
the top of the csproj file in order to get its static variables initialized
before everything else in the final JavaScript. Cool, cool.