wwtlib/URLHelpers.cs: add a mode parameter to URLHelpers.rewrite()
I've had a feeling for a while that I'd need to add this, and this was spurred
by issues with the production beta webclient having problems loading a folder
with URL `../Resource/...`.

I want a mode parameter here because if we load up a WTML file that contains a
relative URL, that URL should be absolute-ified relative to the source of the
WTML file, not the window origin. Right now, relative URLs in WTML files are
not allowed, but I'd like that to change one day. (If/when we implement this,
this function will probably need to acquire a third parameter to give the base
URL to use in the relative-to-something-else case.)