wwtlib/WWTControl.cs: try to avoid the DSS explosion problem
As described in the comment: if we try to render with a zero-size viewport,
which can happen in HTML-land, we'll get NaNs that cause the tile code to
decide it needs to download the highest-resolution tiles available for the
entire viewport. With a default view of the DSS, that leads to the browser
trying to slurp down absolutely massive amounts of data.