azure-pipelines.yml: also publish version tags to a `major.minor` directory
When a pull request is merged to master, it will be uploaded to the
`engine/latest` directory on the Azure Storage. When a version tag is pushed,
it will currently be uploaded to (e.g.) `engine/1.2.3`. Here we add another
stanza to also uploaded tagged release to a "major.minor" directory, e.g.
`engine/1.2`. This way, a web app that wants to get bugfixes can just target
the `1.2` directory and not have to track every individual release. This is of
course assuming a semver model where micro updates do not break anything.

We may need to purge the CDN cache upon releases in order for this workflow to
fully work. It looks like that can be automated as well.