"Hoist" the Vue deps to the top level
If we're going to be including the embed here, it makes sense to centralize
these deps as well.

Ran into a confounding error with the Vue CLI build process at first; would
get a message about `opts` being undefined when processing the custom
`vue.config.js` in `engine-vuex`. It turns out that we need to preserve the
`@vue/cli-plugin-*` packages in the devDependencies of the package.json files.
Without `cli-plugin-typescript` the CLI seems not to know to import the right
TypeScript magic into its webpack config, and various other things break
without the other plugins when they're needed.