fix(engine-types): export a ConstellationFilterInterface
Not a ConstellationFilter class, since that's only actually implemented in the
engine. The engine types will mark that the class implements the interface.

Nominally this should be a breaking change, but this interface hasn't seen the
light of day yet and I want to try out potential automated release processes,
so we'll label it as merely a fix.