feat(engine): make the Wtml class public in C#
As far as I can tell, there is currently no mechanism to obtain a Folder
object *and* register its imagesets with the WWTControl. You can either use
ScriptInterface.LoadImageCollection() and get the imagesets, or make your own
Folder object (and load using LoadFromUrl()) but have no way to register its
imagesets.

The Wtml.GetWtmlFile() function *almost* allows both, but isn't available to
JS because the class is private. This looks like an oversight, so let's just
change that. By also changing GetWtmlFile() to return the Folder that it
creates, we can actually get both things that we want.