feat: work on support for showcasing a single imageset
We expect that a common use case for the embed will be to show a single image
in context. So, we want to have infrastructure to initialize the view right on
that image.

This is actually a bit tricky since in order to do this, we need to know where
to center the view, the zoom level, an appropriate background image, etc. If
we're loading our image from a WTML file, we don't know any of that stuff
until we load the file, so we have to do the setup on the fly.