fix(engine,embed): work on proper initialization of planetary images
The new setup scheme set us up with the wrong zoom factor for viewing things
in planetary mode. It was pretty hard to trace through how the zoom factors
get initialized, but in the end it all seems to be pretty ad-hoc, so I just
special-case the setting. (The default is 360, which make the planet pretty
small in the viewport.)

In the course of this I discovered a nicer way to tell TypeScript about the
Vuex-linked functions in the WWTAwareComponent.

I also needed to alter the API to setupForImageset incompatibly because the
Vuex component setup library seems to not work with functions that take more
than one argument -- the rest are silently dropped.