feat(creator): add UI for creating embeds with AstroPix images
This is kind of clunky, but it at least seems to work. Ideally we'd just get
everything ourselves instead of having the user follow a recipe to get a
ShowImage URL, but I can't see a way to do that without setting up some new
web service APIs -- besides the hassle of, e.g., trying to parse AstroPix HTML
to pull out the ShowImage URL in the client, we're not allowed to even try
because of CORS.

The only thing specific to AstroPix here are the instructions -- really we
just need a ShowImage URL. So, adding additional services should be pretty
easy.