build(astro/tsconfig.json): stop referencing toplevel tsconfig.json
It looks like the package published to NPM will include a reference to
`../tsconfig.json`, making the package un-buildable on its own. (Indeed it's
hard to see how the tooling could be smart enough to do anything about this.)
So I guess we can't actually leverage the monorepo-ness here. Ah well.