ci(azure-pipelines.yml): rework release workflow
Besides being a little intellectually dissatisfying, it didn't work out to
publish beta NPM packages on updates to `master` while non-pre-releases got
published on updates to `release`: because the NPM package releases are how we
export information about new releases to the other deployment processes
(through our `postpublish` scripts), every deployment step would need to have
careful logic to run on either `release` or `master` depending on the
beta-ness. That might be possible to make work, but just seems gnarly.

Fortunately, hopefully, the `--pre-dist-tag` option to `lerna publish` should,
if it behaves as documented, let us upload any prelease packages to the `beta`
channel and therefore gets rid of the original motivation for having the NPM
publish happen on both `master` and `release`, which was to help ensure that
the beta packages land in the `beta` dist-tag.

Basically everything happens at the end of the current pipeline file should
only happen in the `release` branch, and so we should probably do something
smarter than adding Pipelines conditions everywhere. But I think the current
approach should work.