build(graduate.js): attempt to fix release versioning
I ran into an annoying problem when trying to release the `embed` package: I
want the released package to depend on `engine-vuex ^0.1.0`, but the version
of the package within the repo is currently `^0.1.1-beta.0`, and the latter is
viewed as not satisfying the former dependency because of how NPM handles
prereleases specially. This disjunction leads `lerna` to not be able to
properly understand the intra-repo dependency graph, leading to build errors.

Semi-unfortunately, the actual package release mechanics are all handled
inside `lerna`, so we can't do any just-in-time patching or anything.

The solution, I hope, is to just us a `||` clause in the version dependency
specification. Then, well, either version works! I'm not sure if that will
cause other people's packages to potentially pull in beta versions when they
shouldn't, but if we use dist-tags properly I'm pretty sure that will be OK.

We might need to add some further smarts to the script to not sometimes create
dependencies that look like `^0.1.1 || ^0.1.0 || ^0.1.2-beta.0` but we'll
cross that bridge if/when we come to it.