fix(engine-vuex): clean up sync between engine and Vuex state
It was pretty gross to have different mutations for each item in the Vuex
state, and it turns out that you just don't need to separate them out that
way. Now instead of having the Component's render callback do stuff
individually, it just calls an `internalUpdate` function on the store that is
charged with synchronizing everything between Vuex and the underlying WWT
engine.

As part of this, we now properly sync up the background and foreground
imagesets and the foreground opacity. I think.