azure-pipelines.yml: publish engine JS to /$major/, not /$major.$minor/
We're trying to preserve semver and we want users to get the latest
stuff, so it will be much more convenient to not need to update them
every time a minor release comes out. With semver semantics, we expect
our version numbers to evolve as `tiny.large.small`, not
`tiny.small.large`.