Add permissions for WWT embed <iframe>s
Now that we support fullscreening, we should update our iframes to use
the `allowfullscreen` option to make it usable in embeds. We also ask
for a variety of other permissions that we don't currently use, but
might want to use in the future. This way if we gain these capabilities
in the future, preexisting iframes will be able to take advantage of
them.