embed: add loading and tour-play overlays
If we're not in tour-player mode, we jump directly to the "started"
state. One day, we might want to have a start button in these modes as
well, if we can provide a poster of the image and/or pop something up
as the WWT engine JS loads in the background.

Moon loader icon derived from: https://svs.gsfc.nasa.gov/4310 ,
creditable to NASA's Scientific Visualization Studio.