engine: add TourPlayer.LeaveSettingsWhenStopped
Before, when you stopped playing a tour, the temporarily active
"TourSettings" were always cleared. If the tour had a setting like
localHorizonMode active, this could cause a jarring jump in the view.
Here we add a LeaveSettingsWhenStopped option that allows you to make it
so that the tour's settings stay active after it stops (or pauses). This
gives us a way to avoid the jarring jump while keeping the defaul
behavior and API.