engine-vuex(!): add tourStopStartTimes
Add a property giving the start times of the "stops" in the current
tour.

This has become a breaking change since I cannot figure out how to
provide both `loadTour` and a `loadAndPlayTour` as separate Actions that
share a common implementation. So I've removed `loadAndPlayTour` and
replaced it with `loadTour` that takes a `play` boolean argument. To get
the old functionality, use the new action and set `play` to true.

In detail, these now need to be mutation-actions and I think that basic
issue is that the Vuex module decorator does a lot of funky
transformations on the underlying code that make it so that you can't
just delegate to a shared async function.

I'm OK with doing this since I'm quite sure that the embed is the only
user of this API that's out there.