embed: give overlays a z-index of 10 to fix UI controls when paused
On Firefox mobile, when I pause the tour, I can't use any of the UI
controls. This seems to be fixed when I give the elements a z-index
above the main app screen, which is a bit odd since that seems like
something that shouldn't have an effect that depends on the play/pause
state. I think this all might be because pushing the pause button moves
the UI focus somewhere new? But anyway, this seems to fix it.