engine: add Overlay.PrepMultimedia() and use it for audio
More work towards getting audio to work in Safari. We'll need to make
sure to prepare all of our audio tracks in the event handler that
reactions to the user's "play" click.

Here we change the behavior of most methods to use PrepMultimedia()
instead of InitializeTexture(). While InitializeTexture() "ought" to be
called for every frame of playback, note that the music and audio tracks
aren't rendered the same way as other overlays, so that doesn't happen
for them.