engine: fix audio sync when navigating to a master slide
If we navigate tour playback to a master slide, such as slide 0, the
slide's audio will be Play()ed but no Seek() call will occur, since
PlayMasterForCurrent() won't do anything. This means that if the audio
was already playing, we'll resume playing from where it was, not the
beginning as needed. The solution is to make sure that we seek to the
beginning of the audio track(s) when (re)starting the audio in Render().