engine(TourPlay): seek then play, not the opposite
This was causing problems when restarting tour playback from the end --
the "play" command tried to initiate audio playback from the very end of
the stream, so that it instantly finished. Seeking to the beginning of
the stream then didn't do anything. Instead we should seek the
"playhead" to the right place -- offset 0 -- and *then* start playback.