Import the research app
I had hoped that I could keep it in its own repo, but there's going to
be just too much coupling between the app and the engine -- development
in separate repos was not going to be practical. Let's import the app
here so that we can actually make progress on it.