research-app-messages: add a sessionId field to ViewStateMessage
This will allow clients that are talking to multiple apps (e.g.,
the Jupyter ipywidgets implementation) to know which one they're
getting updates from.

This is a breaking change in a certain sense, but only for the research
app implementation -- for clients, it's just a new field they can
access.