research-app: fix cross-origin messaging on Chrome
OK, I think I reminded myself why we were avoiding `event.source
instanceof Window` before. I just ran into a problem with the
"new-and-improved" messaging that came down to this issue: the problem
is specific to cross-origin events on Chrome, which is probably why we
had trouble reproducing it. New documentation tries to record the
reasoning for posterity.