research-app: avoid a startup race condition
The app shouldn't start processing messages until the WWT engine is
fully initialized. Clients are going to have to handle the case that the
app isn't responding to messages anyway, so we don't need to distinguish
between "listening to messages but not yet ready" and "not even
listening to messages".