research-app: add ApplicationStateMessage and wire in HiPS catalog updates
This will be our mechanism for notifying clients when the available list
of HiPS catalogs changes. The general ApplicationStateMessage framework
will also help us synchronize pywwt with the app in general, helping
solve issues with things like recognizing the current foreground and
settings. But I want to get the HiPS catalog data stuff working so we'll
start small.