research-app-messages: fix the FITS fix
This mostly reverts 6b26e24679f67f1cef971a92230a3a6c4149ce07 . I failed
to appreciate that the `mode` field of CreateImageSetLayerMessage should
in fact *not* be optional because that's the only way that we can
distinguish that message from CreateFitsLayerMessage. The real issue was
that the type guard function for the latter should explicitly check that
the message is lacking a mode field, and applyBaseUrlIfApplicable()
should explicitly handle both message types. This should fix up loading
of non-FITS imageset layers now, without breaking URL rewriting of
messages from pywwt.