engine: define some interfaces for settings
I've convinced myself that, in order to be able to wire up Vue(x) user
interface that affects the settings of various WWT objects, we'll need
to mirror those settings in the Vuex state store, because otherwise the
Vue reactivity system can't properly detect and propagate changes. To
help make it so that we can do this mirroring in the face of our
"settings" framework, we extract out interfaces for reading and writing
the settings, so that the same code that modifies the settings of the
WWT objects can also modify "mirror" objects that are wired into the
reactivity framework.