engine-helpers: separate out some settings implementations
For the Vue(x) reactivity work, we need to add a few more bells and
whistles to the settings framework. In particular we add
`extract*Settings()`, `copy*Settings()`, and `*State` classes that serve
as plain-old-data holders for settings values. This is a good reason to
pull out the implementations of the various groups of settings into
separate files.

This commit only implements the changes for Layer and SpreadSheetLayer
settings, to enable my current UI work, but if we like the design we
should apply it to all WWT types that have settings.