research-app: tidy up color/visibility handling
Now that the `Layer.enabled` field is available and we have a way to
work with reactive settings, we can simplify the way that the layer
settings are managed here. We still need the app-specific "visibility"
array to implement the `visibleHipsCatalogs` for now, but that could be
normalized away with a bit more work.