engine: work towards proper FitsProperties setup for FITS imagesets
Non-HiPS tiled FITS imagesets weren't having their LowerCut/UpperCut
settings initialized correctly. The basic reason for this was that,
unlike FITS HiPS, there was no mechanism to set up the FitsProperties
once the first data were loaded. This in turn was the case because
unlike HiPS datasets, there's no `properties` file to download that
specifies these things, and the tile-loading infrastructure didn't have
any hooks to allow notification when the first data come in.

Here we modify the tile-fetching code to add a callback that's triggered
when the level-0 FITS tile is loaded. We then set up FITS imageset
loader to trigger that callback and use a new method to initialize the
FitsProperties from the FITS headers of the level-0 tile.

We also plumb this callback API through more thoroughly so that
`ScriptInterface.AddImageSet()` can now potentially trigger its callback
when these data have loaded, as opposed to always calling its callback
on-the-spot. This is only properly implemented for non-HiPS tiled FITS
right now, but going forward we can improve
`LayerManager.AddImageSetLayerCallback()` to handle other cases more
robustly as well. This way, the research app UI actually sees the
correct cutoff values when it gets notified that the imageset has
loaded.