tests: tidy up README
Tidy up some of the directions a bit, and add `nightwatch` back to the
`package.json` (while maintaining the listing in the toplevel
`package.json` as well) because it seems to be needed on my machine. We
can make it possible to alter the testing environments with commands of
the form `npm run local -- -e firefox` if we remove the `-e default` in
the command specifications: it looks like multiple instances of the `-e`
argument don't override one another.