Add a matrix of options to the BrowserStack configuration.
To do this, I promoted the config files to be TypeScript files, and created
config.ts, which contains some utility functions for adding a matrix of
environments, together with some interfaces for the configuration files
that will hopefully help keep things organized. Everything that we use
now ends up in the dist folder, so I modified the local and bs-local
commands to match this.