engine: correct the ViewCenter computations
I finally sat down and dug into what we're doing here. My initial work
had some serious bugs! We need to account for the rotation of the
coordinate system. Also, in the untiled case it's unfortunately not
possible to get a reliable answer, since we'd need to know the image
dimensions.