engine/wwtlib/TangentTile.cs: organize and tidy coordinate calculations
Get all of the code into the same place and try to make it more
readable, taking care to honor parentheses that might affect
order-of-operations since we can be dealing with high-precision
calculations. The only semantic changes here should be changes of casts
to `double` that are now more internally consistent. In JavaScript-land
this shouldn't affect any results.