Move dependencies into individual packages.
Now that we are using yarn, we can set up the dependencies of each
package in their individual package.json files. This is correct since it
means that published modules have correct dependency specifications in
their metadata, and also fixes commands like `yarn run build` run from
subdirectories. Yarn is smart enough to install all of the deps at the
toplevel when possible, so it's not like we're installing tons of
redundant copies of all of the deps. There's a small downside in that if
we want to, e.g., update TypeScript, we have to bump the version in all
of the package.json files, but that's the correct way to be doing things
anyway.