engine: no builtin image sets in freestanding mode
Also switch this to use the `worldwidetelescope.org` API, rather than
the static asset that we've had for a while. The static asset was
introduced because we didn't have any infrastructure to update the API
in any manageable way. With the introduction of the `wwt-core-catalogs`
framework that issue has been fixed, so now we can re-centralize
everything. This will help with maintainability.