engine: update freestanding mode to take a baseurl for static assets
We don't want to depend on the WWT servers in this mode, but there are
also engine static assets that (1) are easy to host yourself but (2) we
don't want to hardcode the reliance on the `wwtassets.org` server. So,
wire things up so that you tell the engine where to find those assets.