Have the `addImagesetToRepository` method return an imageset
This method now returns that imageset that actually ended up in the
repository. This might be the imageset that the caller wanted to add, or
it might be a preexisting one, if there is already an imageset with the
same URL in the database. The difference might matter if one wishes to
do a name-based lookup of the imageset, since the two equivalent
imagesets might have different name fields.