Add `type: module` specifications to TypeScript modules
We compile our TypeScript to ESM syntax, so we that's what we should do
in the current ecosystem. This doesn't apply to `@wwtelescope/engine`,
which uses UMD module structure (which is compatible with Node/CJS), and
it doesn't seem to work with the Vue components, where it's less of a
priority since we don't really expect other packages to depend on them.

It seems that the current recommendation is to compile to ESM and
potentially transpile to CJS if that's needed. We could look into
providing both CJS and ESM files, which seems increasingly common these
days, but I don't have a sense for much much we need that. The change
here should at least make our files more correct.