engine: change Tile.prepDevice to be a value in esm/render_globals.js
I'm trying to avoid circular dependencies in our module structure,
although they are OK in ESM modules to some extent. Some of these
dependencies are caused by global/static state variables, like
`Tile.prepDevice`. By collecting these variables in modules with
minimal code, we can declare them early in the dependency graph and
avoid the cycles.

This *is* an API change, since the field `wwtlib.Tile.prepDevice` is no
longer being defined or updated. In this particular case, I think that's
OK. We could avoid this by putting a hook mechanism in
`render_globals.js` so that when the prep device is update, a callback
could be invoked that would update `Tile.prepDevice`. This way we could
update the variable under the "old name" without needing
`render_globals.js` to depend on the `tile.js` module.