engine/esm/transpiled.js: export formerly internal symbols
These are/were internal to WWT, but as we break things up into modules,
we'll need to export them.

We're losing a bit of information about privacy here, but in the big
picture, the traditional engine API just didn't have any effort put into
distinguishing public/private types, so we're going to need to
completely revisit this stuff anyway if it's something that we decide
that we want to clean up.