engine-pinia: blah, reorganize the API reference
Re-exported Pinia methods on the WWTAwareComponent weren't getting
documented correctly. So, bite the bullet and move all of the docs to be
associated with `engineStore()` rather than WWTAwareComponent. This is
better in the long run anyway, since Vue-land is pushing people towards
the Composition API.

This causes some new warnings because *some* of the engineStore docs end
up getting attached to the WWTAwareComponent, and the internal links to
renamed properties break in that context, but ... whatever.