docs: try to fix the new images, and other config tweaking
The new images are rendering with URLs pointing to `.../1/...`, rather
than `.../latest/...`, so they're not working in the latest docs. This
seems to come directly out of the base URL specified in the Zola config,
so for now let's just force it to be what I want. This should perhaps be
Cranko-ified, or I should otherwise straighten out my versioning scheme.

Also add protocols to URLs and update Google Analytics IDs.