engine/esm/url_helpers.js: fix HTTPS rewriting corner case
If we were given a URL like `http://mydomain.com:80/` to rewrite in
force-HTTPS mode, we'd change it to `https://mydomain.com:80/`, which
obviously isn't going to work.

In that one case, it makes sense to rewrite to `https://mydomain.com/`
and see what happens. In cases with non-standard port specifications,
that's definitely *not* going to work, so in those cases, we just have
to keep the URL as insecure HTTP and hope the browser doesn't ding us.

This change motivated by an issue in `ShowImage.aspx`, which seems to
have started including the `:80` in the URLs it emits in its redirects,
breaking the Astrometry.Net workflow. I'm not sure why that changed, and
we could try to change it back, but this modification here makes the
rewriting code more correct, and should fully solve the problem. I hope.