engine-pinia: correct loadTour() and overall tour timeline scrubbing
In the port to Pinia, the behavior of the `loadTour()` method was
accidentally broken because it was implemented as a somewhat tricky
`MutationAction` — the state items `tourRunTime` and
`tourStopStartTimes` were not actually getting set, which was causing
the tour timeline scrubber not to work at all. The fix is nice and
straightforward in the Pinia framework.