engine: fix sizing of circle annotations
This is a longstanding bug that I've never gotten around to fixing. The
previous code used a magic constant of 44 to go from the nominal circle
radius in degrees to the number actually used in the shader. In StarHunt
I had realized that this was missing a cos(dec) correction, as well as
a constant prefactor. With the cos(dec) applied, `44 * prefactor` is
almost exactly equal to 180 / pi, and using that number to set the
sizing yields results that look perfect. My guess is that the 44 came
from some manual futzing with a test circle at a dec of 40 degrees,
which would yield that number if you didn't realize that the cos(dec)
term was necessary.